requirements.txt
azure-functions>=1.11.0
azure-ai-textanalytics>=5.3.0
azure-core>=1.24.0
requests>=2.28.0

